chore: Renamed files to match component names
