+++ package.json
@@ -11,9 +11,9 @@
     "@supabase/supabase-js": "latest",
     "class-variance-authority": "latest",
     "classnames": "latest",
-    "clsx": "latest",
+    "clsx": "^2.1.1",
     "embla-carousel-react": "latest",
-    "framer-motion": "latest",
+    "framer-motion": "^11.12.0",
     "gsap": "latest",
     "jquery": "latest",
     "lodash": "latest",
@@ -23,7 +23,7 @@
     "react-dom": "latest",
     "react-router": "^7.0.2",
     "react-spinners": "latest",
-    "tailwind-merge": "latest",
+    "tailwind-merge": "^2.5.5",
     "tailwindcss-animate": "latest",
     "web-vitals": "latest"
   },
