+++ src/components/ContactPage/index.tsx
@@ -2,7 +2,7 @@ import { supabase } from '../../services/supabaseClient';
 import React, { useState, useEffect } from 'react';
 import { useNavigate } from 'react-router';
 
-export default function ContactSection() {
+export default function ContactPage() {
   const navigate = useNavigate();
   const [formData, setFormData] = useState({
     name: '',
