+++ src/components/PortfolioPage/index.tsx
@@ -1,0 +1,18 @@
+import { supabase } from '../../services/supabaseClient';
+import React, { useState, useEffect } from 'react';
+import { useNavigate } from 'react-router';
+
+export default function PortfolioPage() {
+  const navigate = useNavigate();
+
+  
+  return (
+    <div className="flex flex-col items-center justify-between px-4 lg:px-20 gap-8 py-16 pt-24">
+      {/* Contact Form Section */}
+      <div className="w-full lg:w-2/3 flex flex-col items-center justify-center gap-8">
+        
+        
+      </div>
+    </div>
+  );
+}
